[MIFOSX-1012] JUnit Test Cases for Account Transfer
